Die Klasse Location etwas verändert, eine switch- case anweisung eingebaut. MoveRoverTest liefert zwar keine Fehler, aber noch einige Warnings, auf die ich als nächstes eingehen werde.
Die Befehle um den Rover zu steuern müssen noch in die Klasse Rover
implementiert werden.

+++ src/Map.java
@@ -4,7 +4,7 @@ import java.awt.geom.*;
 public class Map
 {
 	int [][] a = new int [5][13]; //ein Array mit 5 Zeilen und 13 Spalten für die Darstellung der Marsoberfläche.
-	public void setLocation (double x, double y) {
+	public void setLocation (int x, int y) {
 		return;
 	}
 }
